Create Foreground Service to register shutdown receiver
Starting with Android 9, receivers listening to ACTION_SHUTDOWN
must be registered through Context.registerReceiver.

Since the receiver needs to be active until the device shuts down,
we have to use a service that stays alive even when the app is in
background or was never launched to begin with.

A foreground service allows us to do just that with the limitation
that a notification has to be visible to the user. As a result
it also entirely replaces the old compatibility mode that used a
notification to keep the app in memory.