Check permissions in both broadcast receivers
In the case that the Do not Disturb permission is revoked while Silent Boot
is enabled, both receivers should not perform any actions, as they will
most likely fail and result in errors.

Users should notice that something is off, as the boot receiver will also
not launch the service. The notification that Silent Boot is active won't
appear as a result.